[fix] Fixed Tracker Highlighting
The Min/Max buttons were causing some weird anomalies with Quest highlighting in that occasionally it would highlight not only the hovered Quest but other items in the Tracker would highlight. So, I removed the highlighting from the Min/Max buttons specifically. Besides, they already get highlighted when you hover over a quest title or an objective. Everything in that "group" gets highlighted so having it also trigger on buttons is somewhat redundant.