[fix] Fixed quests showing as available from hostile NPCs
Curiously this was not that big of a problem, only 6 and 8 quests
resepctively for max level chars on Classic and Cata. Many seem to have
been fixed by overwritting the quest starters in both of the faction
specific corrections. These could now be moved to normal corrections or
entirely removed if the base DB has the correct quest starters (some
don't).